refactored CMS start up - it no longer uses falls back on default content
